feat: Use interactive tokens for all focus styles.
We're also adding a secondary focusOutlineLight style for subtler focus,
useful for text links where a 4px outline is too much.

Right now, we're overwriting this focus style for buttons manually,
to ensure that outlines match the button colour,
but we'll want to make this a little bit more bulletproof in the future.