fix: Hide tab headings. / Visually, these don't make sense to show since the tab itself already provides this context.
We're also switching it from a heading to a tab to ensure consistent heading hierarchy. If we wanted to stick with headings, I might recommend opting for something higher, like an h2, which is more likely to work with the page hierarchy, or pass a prop to declare the heading level.